🚧 Nodes are now fully draggable
Nodes in the Behaviour editor window can be dragged around to be repositioned.

It worries me that I might be allocating a large quantity of JTokens to represent the new values and this could potentially cause GC spikes in the editor whilst dragging.

To avoid this I can conditionally cast the JToken to a JValue and set the JValue.Value property. This will avoid allocations for non-object type fields.