⚡️ Remove generic node instantiation in favour of boilerplate
Removed the overhead of reflection or generic driven instantiation in favour of a boilerplate line of code that instantiates the node instance.

Moving this line of code here removes more mysticism behind the system and could even allow for custom parameters to be passed in to allow for node instances to be reused for multiple nodes.

Finally, multiple nodes could instantiate generic instances of the same node instance.