💥 Changed the connection mapping API
The connection mapping API now uses a ConnectionMapper ref struct (instead of an IGraphConnections). This allows the node to be known to the GraphInstance via the IConnectionCallback methods.

Previously this was being tracked by setting a private field and then invoking the connection on the nodes, and using the last-set node as the "caller".