🚧 Work-in-progress resource tagging and indexing
Basic models for determining all tags that are on a specific resource and all resources with specified tags.

As per #18, tags should be addable as apart of a build post-process step.

I'm battling with whether having multiple ResourceExporters might be appropriate. These are currently responsible for post-processing assets, however making it so there is more than one of them might prove restrictive in the future, as a single ResourceExporter per-resource has a lot more control over the export pipeline.