✨ MVP of "Package resources tags and indexes" (#18)
The ProjectExplorer now uses the the ImportPipeline to import resources from disk.
Added the ProjectResourceImportProcessor base class that is injected into the new ImportPipeline.
The ImportPipeline runs all of the ProjectResourceImportProcessors on the imported assets.
The ProjectResourceImportProcessors can tag resources as they are imported.

The ProjectExplorer and PackageExplorer now contain ITagCollections which can be used to query resources by their tags.
The PackageResource and ProjectResource now contain an IResourceTags that list all of the tags of this resource.